Refactored component and page names
