+++ src/components/Layout/Layout.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import NavBar from '../NavBar'
-import Header from '../Header'
+import NavBar from '../NavBar/NavBar'
+import Header from '../Header/Header'
 
-import './style.scss'
+import './Layout.scss'
 
 export default function index(props) {
     return (
