+++ src/components/NavBarItem/NavBarItem.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { NavLink } from 'react-router-dom'
-import './style.scss'
+import './NavBarItem.scss'
 
 export default function index(props) {
     return (
