+++ src/components/ProfileIcon/ProfileIcon.scss
