+++ src/pages/NotFound/NotFound.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Meta from '../../components/Meta'
+import Meta from '../../components/Meta/Meta'
 
 export default function index() {
     return (
