+++ src/pages/Profile/Profile.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Layout from '../../components/Layout'
-import Meta from '../../components/Meta'
+import Layout from '../../components/Layout/Layout'
+import Meta from '../../components/Meta/Meta'
 
 export default function index() {
     return (
